Replace the `bin/find.sh` utility with `fdfind`
The `bin/find.sh` utility was buggy in some cases and was starting to
accumulate technical debt. Instead of continuing to try to develop a
custom solution to the problem, I removed the utility and replaced it
with the `fdfind` utility.

In addition to supporting the `.gitignore` file, the `fdfind` uility
supports the `.ignore` file used by some other CLI search tools. I am
using this file to ignore binary files.

With this configuration in place, by default, the `fdfind` will only
search text files that are tracked by Git. This is the original problem
I was trying to solve with the `bin/find.sh` utility.

To find binary files (e.g., images) you can pass the `--no-ignore`
option to `fdfind`.