Update the log-in page screenshot
The log-in page was recently redesigned. This commit updates the screenshot to match.

Additionally, I have removed the table of contents. With the addition of the FAQ, this no longer worked as a list of log-in methods. GitBook provides a TOC on the right-hand side of the page that is sufficient for navigation.